moved signup logic into controller
