+++ backend/controllers/api/authController.js
@@ -1,4 +1,4 @@
-const db = require('../db/models');
+const db = require('../../db/models');
 const { User } = db;
 
 // See if user is logged in
