--- backend/controllers/userController.js
@@ -1,17 +1,0 @@
-const db = require('../db/models');
-const { User } = db;
-
-// Not sure if its a thing, but I'll organize these in CRUD
-// Create, find, update, delete
-
-// Create a new user
-
-// Find a user by email
-
-// Find a user by username
-
-// Find a user by ID
-
-// Update a user by (id, or email, or username?)
-
-// Delete a user by (id, or email, or username?)
